+++ curlio/quick/form.hpp
@@ -1,17 +1,17 @@
 #pragma once
 
-#include "../request.hpp"
-
 #include <boost/asio.hpp>
 #include <boost/numeric/conversion/cast.hpp>
+#include <curl/curl.h>
 #include <map>
 #include <sstream>
 #include <string_view>
 
 namespace curlio::quick {
 
-inline std::string construct_form(Request& request,
-                                  const std::map<std::string_view, std::string_view>& parameters) noexcept
+/// Constructs a query parameter list from the given associative container (like `std::map`).
+template<typename Associative_container>
+inline std::string construct_form(CURL* handle, const Associative_container& parameters) noexcept
 {
 	std::stringstream ss;
 	bool first = true;
@@ -21,12 +21,11 @@ inline std::string construct_form(Request& request,
 		}
 		first = false;
 
-		auto escaped =
-		  curl_easy_escape(request.native_handle(), key.data(), boost::numeric_cast<int>(key.size()));
+		auto escaped = curl_easy_escape(handle, key.data(), boost::numeric_cast<int>(key.size()));
 		ss << escaped << '=';
 		curl_free(escaped);
 
-		escaped = curl_easy_escape(request.native_handle(), value.data(), boost::numeric_cast<int>(value.size()));
+		escaped = curl_easy_escape(handle, value.data(), boost::numeric_cast<int>(value.size()));
 		ss << escaped;
 		curl_free(escaped);
 	}