+++ curlio/session.hpp
@@ -1,9 +1,11 @@
 #pragma once
 
+#include "config.hpp"
 #include "detail/curl_share_lock.hpp"
 #include "error.hpp"
 #include "log.hpp"
 #include "request.hpp"
+#include "response.hpp"
 
 #include <boost/asio.hpp>
 #include <curl/curl.h>
@@ -23,7 +25,7 @@ public:
 
 	bool is_valid() const noexcept { return _multi_handle != nullptr; }
 	/// Starts the request. Make sure all data is read and the request is awaited.
-	void start(Request& request);
+	CURLIO_NO_DISCARD std::unique_ptr<Response> start(Request& request);
 	void set_cookie_file(std::string file) { _cookie_file = file; }
 	boost::asio::any_io_executor get_executor() noexcept { return _timer.get_executor(); }
 	Session& operator=(Session&& move) = delete;
@@ -37,11 +39,11 @@ private:
 	std::string _cookie_file;
 	/// All active connections.
 	std::map<curl_socket_t, boost::asio::ip::tcp::socket> _sockets;
+	std::map<detail::Shared_data*, std::shared_ptr<detail::Shared_data>> _active_requests;
 
 	void _async_wait(boost::asio::ip::tcp::socket& socket, boost::asio::socket_base::wait_type type);
 	/// Call Request::_finish() for every finished request and remove it from the multi-handle.
 	void _clean_finished();
-	void _handle_timeout(boost::system::error_code ec);
 	static int _socket_callback(CURL* handle, curl_socket_t socket, int what, void* self_pointer,
 	                            void* socket_pointer);
 	static int _multi_timer_callback(CURLM* multi, long timeout_ms, void* self_pointer);
@@ -77,31 +79,40 @@ inline Session::~Session() noexcept
 	}
 }
 
-inline void Session::start(Request& request)
+inline std::unique_ptr<Response> Session::start(Request& request)
 {
-	if (request._executor) {
+	if (!request.is_valid() || request.is_active()) {
 		throw std::system_error{ Code::request_in_use };
 	}
 
 	CURLIO_DEBUG("Starting request " << &request);
-	const auto easy_handle = request.native_handle();
-	request._executor      = get_executor();
-	curl_easy_setopt(easy_handle, CURLOPT_OPENSOCKETFUNCTION, &Session::_open_socket);
-	curl_easy_setopt(easy_handle, CURLOPT_OPENSOCKETDATA, this);
-	curl_easy_setopt(easy_handle, CURLOPT_CLOSESOCKETFUNCTION, &Session::_close_socket);
-	curl_easy_setopt(easy_handle, CURLOPT_CLOSESOCKETDATA, this);
-	curl_easy_setopt(easy_handle, CURLOPT_COOKIEFILE, _cookie_file.c_str());
+	auto data         = std::move(request._owner);
+	const auto handle = data->handle;
+	data->executor    = get_executor();
+
+	curl_easy_setopt(handle, CURLOPT_OPENSOCKETFUNCTION, &Session::_open_socket);
+	curl_easy_setopt(handle, CURLOPT_OPENSOCKETDATA, this);
+	curl_easy_setopt(handle, CURLOPT_CLOSESOCKETFUNCTION, &Session::_close_socket);
+	curl_easy_setopt(handle, CURLOPT_CLOSESOCKETDATA, this);
+	curl_easy_setopt(handle, CURLOPT_PRIVATE, data.get());
+	curl_easy_setopt(handle, CURLOPT_COOKIEFILE, _cookie_file.c_str());
 	if (!_cookie_file.empty()) {
 		CURLIO_DEBUG("Setting cookie file '" << _cookie_file << "' for request " << &request);
-		curl_easy_setopt(easy_handle, CURLOPT_COOKIEJAR, _cookie_file.c_str());
+		curl_easy_setopt(handle, CURLOPT_COOKIEJAR, _cookie_file.c_str());
 	}
-	curl_easy_setopt(easy_handle, CURLOPT_SHARE, _share_handle);
+	curl_easy_setopt(handle, CURLOPT_SHARE, _share_handle);
 
-	curl_multi_add_handle(_multi_handle, easy_handle);
+	std::unique_ptr<Response> response{ new Response{ data } };
+	_active_requests.insert({ data.get(), std::move(data) });
 
-	int running = 0;
-	curl_multi_socket_action(_multi_handle, CURL_SOCKET_TIMEOUT, 0, &running);
-	CURLIO_DEBUG("Kick-off timeout. Running=" << running);
+	curl_multi_add_handle(_multi_handle, handle);
+	int still_running = 0;
+	// curl_multi_socket_action(_multi_handle, CURL_SOCKET_TIMEOUT, 0, &running);
+	curl_multi_perform(_multi_handle, &still_running);
+	CURLIO_DEBUG("Kickstart everything. Running=" << still_running);
+	_clean_finished();
+
+	return response;
 }
 
 inline void Session::_async_wait(boost::asio::ip::tcp::socket& socket,
@@ -112,8 +123,12 @@ inline void Session::_async_wait(boost::asio::ip::tcp::socket& socket,
 		if (!ec) {
 			const auto handle = socket.native_handle();
 			int still_running = 0;
-			curl_multi_socket_action(_multi_handle, handle, 0, &still_running);
-			// _clean_finished();
+			const auto code   = curl_multi_socket_action(_multi_handle, handle, 0, &still_running);
+			if (code != CURLM_OK) {
+				CURLIO_ERROR("====================" << curl_multi_strerror(code));
+			}
+			_clean_finished();
+
 			if (still_running <= 0) {
 				_timer.cancel();
 			}
@@ -132,25 +147,13 @@ inline void Session::_clean_finished()
 	int left         = 0;
 	while ((message = curl_multi_info_read(_multi_handle, &left))) {
 		if (message->msg == CURLMSG_DONE) {
-			Request* request = nullptr;
-			curl_easy_getinfo(message->easy_handle, CURLINFO_PRIVATE, &request);
-			CURLIO_DEBUG("Request " << request << " is done");
-			if (request != nullptr) {
-				request->_finish();
+			detail::Shared_data* data = nullptr;
+			if (curl_easy_getinfo(message->easy_handle, CURLINFO_PRIVATE, &data) == CURLE_OK && data != nullptr) {
+				CURLIO_DEBUG("Request " << data << " is done");
+				_active_requests.erase(data);
 			}
 			curl_multi_remove_handle(_multi_handle, message->easy_handle);
 		}
-	}
-}
-
-inline void Session::_handle_timeout(boost::system::error_code ec)
-{
-	CURLIO_TRACE("Socket timeouted with ec=" << ec);
-	if (!ec) {
-		int still_running = 0;
-		curl_multi_socket_action(_multi_handle, CURL_SOCKET_TIMEOUT, 0, &still_running);
-		CURLIO_DEBUG(still_running << " requests still running");
-		_clean_finished();
 	}
 }
 
@@ -165,10 +168,10 @@ inline int Session::_socket_callback(CURL* handle, curl_socket_t socket, int wha
 
 	switch (what) {
 	case CURL_POLL_IN: self->_async_wait(it->second, boost::asio::socket_base::wait_read); break;
-	case CURL_POLL_OUT: self->_async_wait(it->second, boost::asio::socket_base::wait_write); break;
+	case CURL_POLL_OUT: /* self->_async_wait(it->second, boost::asio::socket_base::wait_write); */ break;
 	case CURL_POLL_INOUT: {
 		self->_async_wait(it->second, boost::asio::socket_base::wait_read);
-		self->_async_wait(it->second, boost::asio::socket_base::wait_write);
+		// self->_async_wait(it->second, boost::asio::socket_base::wait_write);
 		break;
 	}
 	case CURL_POLL_REMOVE: it->second.cancel(); break;
@@ -181,12 +184,21 @@ inline int Session::_multi_timer_callback(CURLM* multi, long timeout_ms, void* s
 	CURLIO_DEBUG("Starting timeout with " << timeout_ms << " ms");
 	const auto self = static_cast<Session*>(self_pointer);
 
-	self->_timer.expires_from_now(std::chrono::milliseconds{ timeout_ms });
-	if (timeout_ms > 0) {
-		self->_timer.async_wait([self](boost::system::error_code ec) { return self->_handle_timeout(ec); });
-	} else {
-		self->_handle_timeout({});
-	}
+	self->_timer.expires_after(std::chrono::milliseconds{ timeout_ms });
+	self->_timer.async_wait([self](boost::system::error_code ec) {
+		CURLIO_TRACE("Timeout with ec=" << ec.what());
+		if (!ec) {
+			int still_running = 0;
+			const auto code = curl_multi_socket_action(self->_multi_handle, CURL_SOCKET_TIMEOUT, 0, &still_running);
+			if (code == CURLM_OK) {
+				CURLIO_DEBUG(still_running << " requests still running");
+				self->_clean_finished();
+			} else {
+				CURLIO_ERROR("====================" << curl_multi_strerror(code));
+			}
+		}
+	});
+
 	return 0;
 }